pyproject.toml
[project]
name = "ProjectQCDashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dash-bootstrap-components>=2.0.4",
    "flask==3.1.2",
    "gunicorn==23.0.0",
    "numpy==2.3.2",
    "pandas==2.3.2",
    "plotly==6.1.2",
    "pyyaml>=6.0.3",
    "watchdog==6.0.0",
]
